Refactor and move bids_inject_sidecar/bids_media to separate reprostim.bids package.
