+++ src/reprostim/bids/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-FileCopyrightText: 2020-2026 ReproNim Team <info@repronim.org>
+#
+# SPDX-License-Identifier: MIT
