+++ src/reprostim/bids/inject_sidecar.py
